This patch provided by Jay Kulpinski of General Dynamics Advanced Information Systems.
Description from provider: The issue is that GET() uses the AS field in the
status word to compute the physical address of the desired word. If you
update the status word before fetching the new value of the instruction
counter, the IC will be fetched using the new address space instead of the
old.